1.1.2
- Changed default p_Huber to be more appropriate
- Fixed bug in that backward search was not returning NULL when trying
to remove the last variable available.
- Added print of number of unique observations and
categories/time=points in longitudinal_folds function as this is pretty
useful
- Added option to set min and max of outcome before running LEGIT or
IMLEGIT so that predictions are never outside the range. Considering
that glm() can't do truncated continuous distributions, this should
already be implemented in glm().
- Added Genetic algorithm variable selection, works much better than
other variable selection functions but it's very slow and need a lot of
parameter tweaking.